Move sort-button into history-page
