+++ src/components/history-page/history-toolbar/history-toolbar.tsx
@@ -6,7 +6,7 @@ import { useSelector } from 'react-redux'
 import { ApplicationState } from '../../../redux'
 import { ForkAwesomeIcon } from '../../common/fork-awesome/fork-awesome-icon'
 import { ShowIf } from '../../common/show-if/show-if'
-import { SortButton, SortModeEnum } from '../../common/sort-button/sort-button'
+import { SortButton, SortModeEnum } from '../sort-button/sort-button'
 import { HistoryEntry } from '../history-page'
 import { ClearHistoryButton } from './clear-history-button'
 import { ExportHistoryButton } from './export-history-button'
