+++ src/components/history-page/sort-button/sort-button.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { ButtonProps } from 'react-bootstrap'
-import { IconName } from '../fork-awesome/fork-awesome-icon'
-import { IconButton } from '../icon-button/icon-button'
+import { IconName } from '../../common/fork-awesome/fork-awesome-icon'
+import { IconButton } from '../../common/icon-button/icon-button'
 
 export enum SortModeEnum {
     up = 1,
